Update README.md
The project's primary emphasis is on cleaning and pre-processing all of the photos, including coding to remove duplicates, convert from PNG to JPEG format, normalise and scale the images, and extract features based on HOG values. The classifiers will be trained using the extracted features, and then evaluated using 10 rounds of cross validation. In order for the trained classifiers to determine which class this picture belongs to, the hyperparameters of the classifiers will be fine-tuned and the features from the test dataset will be input into the model. The remainder of the paper will show and assess the findings obtained from the classifiers and models, provide suggestions for the best models and classifiers based on the analysis, and provide ideas for increasing the efficiency of the classifiers and models. Support vector machine, decision tree, Logistic regression, K-Nearest neighbour, Convolutional Neural Network, and full connected neural network are just some of the linear and non-linear classifiers used in this paper.